Implemented usbcomm layer to print and read the USB serial port
Implements an init buffer that buffer the init messages and read them
back when the serial port is opened. This permits to be able to read about
sensor initialization and other startup messages via USB (the normal
way would be to reset the board to see the messages again but this is not
working with USB as it would disconnects the USB serial port on reset).